Add initial experimentation file.
This is just a file to experiment with in order to get a better understanding of how RMarkdown can create HTML infographics.  It isn't meant to recreate any infographic or even to serve as template in order to do so.  The principles discovered by playing with this file can be used to help generate a template for infographics, but the code itself shouldn't be used directly for the template.